requirements.txt
# Core LLM & Embeddings
openai>=1.40.0
anthropic>=0.34.0

# Vector Database
chromadb>=0.5.0

# Document Processing
pymupdf>=1.24.0
pillow>=10.4.0
python-docx>=1.1.0

# API Framework
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# UI
streamlit>=1.38.0
streamlit-chat>=0.1.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
httpx>=0.27.0
tenacity>=8.5.0
rich>=13.7.0
tiktoken>=0.7.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Dev
black>=24.8.0
ruff>=0.6.0
